[coord] Share the test wallet fixture
send.rs built one inline; the next test module needs the same thing.
